Featured Post Widget 3.6
git-svn-id: https://plugins.svn.wordpress.org/post-feature-widget/trunk@885587 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ postfeature.php
@@ -3,7 +3,7 @@
 Plugin Name: Featured Post Widget
 Plugin URI: http://wasistlos.waldemarstoffel.com/plugins-fur-wordpress/featured-post-widget
 Description: Featured Post Widget is yet another plugin to make your blog a bit more newspaper-like. Just by choosing a post from a dropdown, you can put it in the 'featured' area and display thumbnail, headline, excerpt or all three of them (if available) in the fully customizable widget.
-Version: 3.5
+Version: 3.6
 Author: Waldemar Stoffel
 Author URI: http://www.waldemarstoffel.com
 License: GPL3
@@ -30,74 +30,63 @@ Text Domain: postfeature
 
 /* Stop direct call */
 
-if(preg_match('#' . basename(__FILE__) . '#', $_SERVER['PHP_SELF'])) die('Sorry, you don\'t have direct access to this page.');
+defined('ABSPATH') OR exit;
 
-/* load classes */
+if (!defined('PF_PATH')) define( 'PF_PATH', plugin_dir_path(__FILE__) );
+if (!defined('PF_BASE')) define( 'PF_BASE', plugin_basename(__FILE__) );
 
-define( 'PFW_PATH', plugin_dir_path(__FILE__) );
+# loading the framework
+if (!class_exists('A5_Image')) require_once PF_PATH.'class-lib/A5_ImageClass.php';
+if (!class_exists('A5_Excerpt')) require_once PF_PATH.'class-lib/A5_ExcerptClass.php';
+if (!class_exists('A5_FormField')) require_once PF_PATH.'class-lib/A5_FormFieldClass.php';
+if (!class_exists('A5_OptionPage')) require_once PF_PATH.'class-lib/A5_OptionPageClass.php';
+if (!class_exists('A5_DynamicFiles')) require_once PF_PATH.'class-lib/A5_DynamicFileClass.php';
 
-if (!class_exists('A5_Image')) require_once PFW_PATH.'class-lib/A5_ImageClass.php';
-if (!class_exists('A5_Excerpt')) require_once PFW_PATH.'class-lib/A5_ExcerptClass.php';
-if (!class_exists('A5_FormField')) require_once PFW_PATH.'class-lib/A5_FormFieldClass.php';
-if (!class_exists('Featured_Post_Widget')) require_once PFW_PATH.'class-lib/FP_WidgetClass.php';
-if (!class_exists('A5_DynamicCSS')) :
-
-	require_once PFW_PATH.'class-lib/A5_DynamicCSSClass.php';
-	
-	$dynamic_css = new A5_DynamicCSS;
-	
-endif;
+#loading plugin specific classes
+if (!class_exists('FP_Admin')) require_once PF_PATH.'class-lib/FP_AdminClass.php';
+if (!class_exists('FP_DynamicCSS')) require_once PF_PATH.'class-lib/FP_DynamicCSSClass.php';
+if (!class_exists('Featured_Post_Widget')) require_once PF_PATH.'class-lib/FP_WidgetClass.php';
 
 class PostFeaturePlugin {
 	
 	const language_file = 'postfeature';
 
+	static $options;
+	
 	function __construct() {
-
-		/* import laguage files */
 		
 		load_plugin_textdomain(self::language_file, false , basename(dirname(__FILE__)).'/languages');
 		
-		add_action('admin_enqueue_scripts', array($this, 'register_scripts'));
-		add_filter('plugin_row_meta', array($this, 'register_links'),10,2);
-		register_activation_hook(__FILE__, array($this, 'install'));
-		register_deactivation_hook(__FILE__, array($this, 'uninstall'));
+		add_action('admin_enqueue_scripts', array(&$this, 'enqueue_scripts'));
 		
-		// attach CSS and write your name in the comments
+		add_filter('plugin_row_meta', array(&$this, 'register_links'), 10, 2);	
+		add_filter( 'plugin_action_links', array(&$this, 'plugin_action_links'), 10, 2 );
+				
+		register_activation_hook(  __FILE__, array(&$this, '_install') );
+		register_deactivation_hook(  __FILE__, array(&$this, '_uninstall') );
 		
-		$eol = "\r\n";
-		$tab = "\t";
+		self::$options = get_option('postfeature_cache');
 		
-		A5_DynamicCSS::$styles .= $eol.'/* CSS portion of the Featured Post Widget */'.$eol.$eol;
+		if (false != self::$options) : 
 		
-		A5_DynamicCSS::$styles.='div[id^="featured_post_widget"].widget_featured_post_widget img {'.$eol.$tab.'height: auto;'.$eol.$tab.'max-width: 100%;'.$eol.'}'.$eol;
+			delete_option('postfeature_cache');
+			
+			$this->_install();
+			
+		endif;
 		
-		A5_DynamicCSS::$styles.='div[id^="featured_post_widget"].widget_featured_post_widget {'.$eol.$tab.'-moz-hyphens: auto;'.$eol.$tab.'-o-hyphens: auto;'.$eol.$tab.'-webkit-hyphens: auto;'.$eol.$tab.'-ms-hyphens: auto;'.$eol.$tab.'hyphens: auto; '.$eol.'}'.$eol;
+		self::$options = get_option('pf_options');
 		
-	}
-	
-	function install() {
+		$FP_DynamicCSS = new FP_DynamicCSS;
+		$FP_Admin = new FP_Admin;
 		
-		$default = array(
-			'tags' => array(),
-			'sizes' => array()
-		);
-	
-		add_option('postfeature_cache', $default);
-	
-	}
-	
-	function uninstall() {
-	
-		delete_option('postfeature_cache');
-	
 	}
 	
 	/* attach JavaScript file for textarea reszing */
 	
-	function register_scripts($hook) {
+	function enqueue_scripts($hook) {
 		
-		if ($hook != 'widgets.php') return;
+		if ($hook != 'settings_page_featured-post-settings' && $hook != 'widgets.php') return;
 		
 		wp_register_script('ta-expander-script', plugins_url('ta-expander.js', __FILE__), array('jquery'), '2.0', true);
 		wp_enqueue_script('ta-expander-script');
@@ -109,8 +98,7 @@ class PostFeaturePlugin {
 	
 	function register_links($links, $file) {
 		
-		$base = plugin_basename(__FILE__);
-		if ($file == $base) :
+		if ($file == PF_BASE) :
 		
 			$links[] = '<a href="http://wordpress.org/extend/plugins/post-feature-widget/faq/" target="_blank">'.__('FAQ', self::language_file).'</a>';
 			$links[] = '<a href="https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=D8AVGNDYYUNA2" target="_blank">'.__('Donate', self::language_file).'</a>';
@@ -118,6 +106,35 @@ class PostFeaturePlugin {
 		endif;
 		
 		return $links;
+	
+	}
+	
+	function plugin_action_links( $links, $file ) {
+		
+		if ($file == PF_BASE) array_unshift($links, '<a href="'.admin_url( 'options-general.php?page=featured-post-settings' ).'">'.__('Settings', self::language_file).'</a>');
+	
+		return $links;
+	
+	}
+	
+	// Creating default options on activation
+	
+	function _install() {
+		
+		$default = array(
+			'cache' => array(),
+			'inline' => false
+		);
+	
+		add_option('pf_options', $default);
+	
+	}
+	
+	// Cleaning on deactivation
+	
+	function _uninstall() {
+	
+		delete_option('pf_options');
 	
 	}
